extract SimpleCardGame to src/game
