+++ src/game/index.ts
@@ -2,4 +2,4 @@ export { Card, Rank, Suit } from './Card';
 export { Deck } from './Deck';
 export { Game } from './Game';
 export { Player } from './Player';
-
+export { SimpleCardGame } from './SimpleCardGame';
